requirements.txt
asgiref==3.7.2
dj-database-url==0.5.0
Django==4.2.10
django-allauth==0.57.2
django-summernote==0.8.20.0
djangorestframework==3.14.0
oauthlib==3.2.2
psycopg2==2.9.9
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests-oauthlib==1.3.1
sqlparse==0.4.4
